refactor: split main logic into modules
